requirements.txt
together>=1.2.0
pyaudio>=0.2.14
python-dotenv>=1.0.0
